refactor(import-sp): split engine/import-engine.ts by phase
Bean: ps-oyu5
