+++ packages/import-sp/src/engine/import-engine.ts
@@ -30,194 +30,36 @@ import {
   completeCollection,
   emptyCheckpointState,
   markRealPrivacyBucketsMapped,
-  resumeStartCollection,
-  type AdvanceDelta,
-  type ImportRunResult,
 } from "@pluralscape/import-core";
 
-import { CHECKPOINT_CHUNK_SIZE } from "../import-sp.constants.js";
-import { synthesizeLegacyBuckets, type MappedPrivacyBucket } from "../mappers/bucket.mapper.js";
-import { createMappingContext, type MappingContext } from "../mappers/context.js";
+import { createMappingContext } from "../mappers/context.js";
 
 import { DEPENDENCY_ORDER } from "./dependency-order.js";
 import { classifyError, isFatalError, ResumeCutoffNotFoundError } from "./engine-errors.js";
 import { collectionToEntityType, entityTypeToCollection } from "./entity-type-map.js";
 import { MAPPER_DISPATCH } from "./mapper-dispatch.js";
+import {
+  buildPersistableEntity,
+  CHECKPOINT_CHUNK_SIZE,
+  delta,
+  indexOfResumeCollection,
+  isAborted,
+  KNOWN_DEPENDENCY_ORDER_SET,
+  makeAbortedResult,
+  makeCompletedResult,
+  persistSynthesizedBuckets,
+} from "./orchestrator-helpers.js";
 
-import type { PersistableEntity, Persister } from "../persistence/persister.types.js";
-import type { ImportDataSource } from "../sources/source.types.js";
 import type { SpCollectionName } from "../sources/sp-collections.js";
-import type {
-  ImportAvatarMode,
-  ImportCheckpointState,
-  ImportCollectionType,
-  ImportError,
-} from "@pluralscape/types";
+import type { ImportCheckpointState, ImportError } from "@pluralscape/types";
 
-/**
- * Check whether the signal has been aborted. Isolated into a function so
- * TypeScript's control-flow narrowing cannot eliminate the second check
- * inside the document loop -- `AbortSignal.aborted` is mutable.
- */
-function isAborted(signal: AbortSignal | undefined): boolean {
-  return signal?.aborted === true;
-}
-
-/** Build a single-document `AdvanceDelta` for one of the four terminal outcomes. */
-function delta(kind: "imported" | "updated" | "skipped" | "failed"): AdvanceDelta {
-  return {
-    imported: kind === "imported" ? 1 : 0,
-    updated: kind === "updated" ? 1 : 0,
-    skipped: kind === "skipped" ? 1 : 0,
-    failed: kind === "failed" ? 1 : 0,
-    total: 1,
-  };
-}
-
-/**
- * Narrow a dispatch result into a {@link PersistableEntity} variant at the
- * upsert boundary.
- *
- * Invariant: `MAPPER_DISPATCH` guarantees that for each collection, the
- * mapper returns a payload whose shape matches the `Mapped<Entity>` type
- * bound to the corresponding `entityType` variant of `PersistableEntity`.
- * We therefore construct the variant with a single controlled cast —
- * avoiding widening the persister's input type or threading a generic
- * through the dispatch table — and the TypeScript compiler enforces the
- * shape of every consumer of the narrowed result. A runtime guard rejects
- * primitives and null payloads so misrouted dispatch entries surface as a
- * visible error rather than a silent cast.
- */
-export function buildPersistableEntity(
-  entityType: ImportCollectionType,
-  sourceEntityId: string,
-  payload: unknown,
-): PersistableEntity {
-  if (payload === null || typeof payload !== "object") {
-    throw new Error(
-      `mapper for ${entityType} returned non-object payload (${typeof payload}); dispatch table may be misrouted`,
-    );
-  }
-  return {
-    entityType,
-    sourceEntityId,
-    source: "simply-plural",
-    payload,
-  } as PersistableEntity;
-}
-
-function aborted(
-  state: ImportCheckpointState,
-  ctx: MappingContext,
-  errors: readonly ImportError[],
-): ImportRunResult {
-  return { finalState: state, warnings: ctx.warnings, errors, outcome: "aborted" };
-}
-
-function completed(
-  state: ImportCheckpointState,
-  ctx: MappingContext,
-  errors: readonly ImportError[],
-): ImportRunResult {
-  return { finalState: state, warnings: ctx.warnings, errors, outcome: "completed" };
-}
-
-export interface RunImportArgs {
-  readonly source: ImportDataSource;
-  readonly persister: Persister;
-  readonly initialCheckpoint?: ImportCheckpointState;
-  readonly options: {
-    readonly selectedCategories: Partial<Record<ImportCollectionType, boolean>>;
-    readonly avatarMode: ImportAvatarMode;
-  };
-  readonly onProgress: (state: ImportCheckpointState) => Promise<void>;
-  readonly abortSignal?: AbortSignal;
-}
-
-// Re-export run-outcome types so SP consumers import from one package.
 export type { ImportRunOutcome, ImportRunResult } from "@pluralscape/import-core";
+export { buildPersistableEntity } from "./orchestrator-helpers.js";
+export type { RunImportArgs } from "./orchestrator-helpers.js";
 
-/**
- * Find the index in `DEPENDENCY_ORDER` of the collection corresponding to a
- * resumed entity type. Resumes are stored as `ImportEntityType`; we walk back
- * to the collection name to know where in the iteration order to start.
- */
-function indexOfResumeCollection(state: ImportCheckpointState): number {
-  const entityType = resumeStartCollection(state);
-  const collection = entityTypeToCollection(entityType);
-  return DEPENDENCY_ORDER.indexOf(collection);
-}
-
-/**
- * Synthesize the three legacy privacy buckets, persist them, and register
- * each in the translation table so member mappers can resolve their
- * `synthetic:*` references. Called only when the source has no
- * `privacyBuckets` collection.
- *
- * Returns the accumulated {@link AdvanceDelta} across all three buckets so
- * the caller can thread it through {@link advanceWithinCollection} and then
- * call {@link completeCollection} before entering the member loop.
- */
-async function persistSynthesizedBuckets(
-  persister: Persister,
-  ctx: MappingContext,
-  errors: ImportError[],
-): Promise<{ delta: AdvanceDelta; aborted: boolean; lastSourceId: string | null }> {
-  const synthesized = synthesizeLegacyBuckets();
-  let importedCount = 0;
-  let updatedCount = 0;
-  let skippedCount = 0;
-  let failedCount = 0;
-  let lastSourceId: string | null = null;
-  for (const bucket of synthesized) {
-    const payload: MappedPrivacyBucket = {
-      encrypted: {
-        name: bucket.name,
-        description: bucket.description,
-      },
-    };
-    try {
-      const result = await persister.upsertEntity({
-        entityType: "privacy-bucket",
-        sourceEntityId: bucket.syntheticSourceId,
-        source: "simply-plural",
-        payload,
-      });
-      ctx.register("privacy-bucket", bucket.syntheticSourceId, result.pluralscapeEntityId);
-      if (result.action === "created") importedCount += 1;
-      else if (result.action === "updated") updatedCount += 1;
-      else skippedCount += 1;
-      lastSourceId = bucket.syntheticSourceId;
-    } catch (thrown) {
-      const error = classifyError(thrown, {
-        entityType: "privacy-bucket",
-        entityId: bucket.syntheticSourceId,
-      });
-      errors.push(error);
-      await persister.recordError(error);
-      failedCount += 1;
-      if (isFatalError(error)) {
-        return { delta: buildDelta(), aborted: true, lastSourceId };
-      }
-    }
-  }
-  return { delta: buildDelta(), aborted: false, lastSourceId };
-
-  function buildDelta(): AdvanceDelta {
-    return {
-      imported: importedCount,
-      updated: updatedCount,
-      skipped: skippedCount,
-      failed: failedCount,
-      total: importedCount + updatedCount + skippedCount + failedCount,
-    };
-  }
-}
-
-/** Set of SP collection names the engine iterates. Computed once per run. */
-const KNOWN_DEPENDENCY_ORDER_SET = new Set<string>(DEPENDENCY_ORDER);
-
-export async function runImport(args: RunImportArgs): Promise<ImportRunResult> {
+export async function runImport(
+  args: import("./orchestrator-helpers.js").RunImportArgs,
+): Promise<import("@pluralscape/import-core").ImportRunResult> {
   const { source, persister, options, onProgress } = args;
   const ctx = createMappingContext({ sourceMode: source.mode });
   const errors: ImportError[] = [];
@@ -318,7 +160,7 @@ export async function runImport(args: RunImportArgs): Promise<ImportRunResult> {
       if (collection === "members" && privacyBucketsMapped === 0) {
         const synth = await persistSynthesizedBuckets(persister, ctx, errors);
         if (synth.aborted) {
-          return aborted(state, ctx, errors);
+          return makeAbortedResult(state, ctx, errors);
         }
         // Advance checkpoint totals for the synthesized buckets, mark the
         // privacy-bucket collection complete, then flush and report progress
@@ -440,7 +282,7 @@ export async function runImport(args: RunImportArgs): Promise<ImportRunResult> {
                 // re-synthesize legacy buckets when the source had real ones.
                 state = markRealPrivacyBucketsMapped(state);
               }
-              let upsertDelta: AdvanceDelta;
+              let upsertDelta: import("@pluralscape/import-core").AdvanceDelta;
               switch (upsert.action) {
                 case "created":
                   upsertDelta = delta("imported");
@@ -513,7 +355,7 @@ export async function runImport(args: RunImportArgs): Promise<ImportRunResult> {
             };
         errors.push(error);
         await persister.recordError(error);
-        return aborted(state, ctx, errors);
+        return makeAbortedResult(state, ctx, errors);
       }
 
       // Resume cutoff sanity check: if we were resuming mid-collection and
@@ -528,11 +370,11 @@ export async function runImport(args: RunImportArgs): Promise<ImportRunResult> {
         );
         errors.push(cutoffError);
         await persister.recordError(cutoffError);
-        return aborted(state, ctx, errors);
+        return makeAbortedResult(state, ctx, errors);
       }
 
       if (collectionAborted) {
-        return aborted(state, ctx, errors);
+        return makeAbortedResult(state, ctx, errors);
       }
 
       // Collection finished cleanly — advance state and flush+report.
@@ -547,7 +389,7 @@ export async function runImport(args: RunImportArgs): Promise<ImportRunResult> {
       }
     }
 
-    return completed(state, ctx, errors);
+    return makeCompletedResult(state, ctx, errors);
   } finally {
     try {
       await source.close();
